// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Branches = {
  TURKCE: 'TURKCE',
  MATEMATIK: 'MATEMATIK',
  INGILIZCE: 'INGILIZCE',
  BEDEN_EGITIMI: 'BEDEN_EGITIMI',
  RESIM: 'RESIM',
  MUZIK: 'MUZIK',
  HAYAT_BILGISI: 'HAYAT_BILGISI',
  FEN_BILGISI: 'FEN_BILGISI',
  SATRANC: 'SATRANC'
} as const

export type Branches = (typeof Branches)[keyof typeof Branches]


export const Grades = {
  ELEMENTARY: 'ELEMENTARY',
  MIDDLE_HIGH: 'MIDDLE_HIGH'
} as const

export type Grades = (typeof Grades)[keyof typeof Grades]


export const Days = {
  MONDAY: 'MONDAY',
  TUESDAY: 'TUESDAY',
  WEDNESDAY: 'WEDNESDAY',
  THURSDAY: 'THURSDAY',
  FRIDAY: 'FRIDAY'
} as const

export type Days = (typeof Days)[keyof typeof Days]